backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
python-multipart>=0.0.6
websockets>=12.0
httpx>=0.27.0
google-genai>=1.0.0
chromadb>=0.4.0
sse-starlette>=1.6.0